refactor: move dpd NAT management to instance_network
